requirements.txt
# Core FastAPI and web framework dependencies
fastapi
uvicorn[standard]
python-multipart

# Database dependencies
psycopg2-binary
sqlalchemy

# Authentication and security
PyJWT

# Optimization and scheduling
ortools

# HTTP requests and utilities
requests

# Additional dependencies for production deployment
gunicorn
python-dotenv
# uvicorn app:app --host 0.0.0.0 --port 8000 --workers 4
